Add files via upload
The "Instagram Quiz Game" is a fun and interactive quiz that tests your knowledge of Instagram. Challenge yourself with a series of questions related to the popular social media platform. Do you know when Instagram was launched, or how many photos are posted daily? Test your Instagram IQ and see if you can answer all the questions correctly.